dmks: use real location or /extramodules
For some reason---probably me having misread some documentation a long
time ago---we were suffixing the location for all modules with /misc.

	DEST_MODULE_LOCATION[#]=/kernel/drivers/hwmon/misc

That appears to be the cause of the issue reported to our AUR package
where modules for old kernels remain in the filesystem after they are
uninstalled.[^1]

DMKS will move pre-existing modules during dkms install, and will
restore them during dkms remove.  For this, DEST_MODULE_LOCATION is
used, and it seems that our setting was causing DKMS to restore the
original module to the wrong path.  Then, when the platform package
manager would uninstall the old kernel, the unknown (to it) file would
be kept.

Use the real path for modules that replace existing ones in mainline
(nzxt-kraken2), or soon to be there (nzxt-smart2).  For the other
modules, use /extramodules.

[^1]: https://aur.archlinux.org/packages/liquidtux-dkms-git/#comment-841782